fix(spiders): remove nigeria_portal_base as only used in gombe_state now
